+++ events.go
@@ -4,10 +4,12 @@ type Event struct {
 	eventType string
 	msg       string
 	err       error
+	sender    User
+	room      Room
 }
 
-func NewEvent(eventType string, msg string, err error) Event {
-	return Event{eventType: eventType, msg: msg, err: err}
+func NewEvent(eventType string, msg string, err error, sender User, room Room) Event {
+	return Event{eventType: eventType, msg: msg, err: err, sender: sender, room: room}
 }
 
 func (e Event) Type() string {
@@ -20,4 +22,12 @@ func (e Event) Message() string {
 
 func (e Event) Error() error {
 	return e.err
+}
+
+func (e Event) Sender() User {
+	return e.sender
+}
+
+func (e Event) Room() Room {
+	return e.room
 }