+++ room.go
@@ -1,0 +1,6 @@
+package dosbot
+
+type Room interface {
+	Name() string
+	ID() interface{}
+}